Missed a closing parenthesis, other minor nitpicks (#175)
Missed a closing parenthesis and changed 'Fedora release' to 'Fedora project'

Added some dollar signs to the start of commands to indicate that they're shell commands and that they should be run as a non-privileged user.  Commands that need to be run as root have hash signs at the start of the line.