Fix input_paths globbing when search paths are defined
`glob.glob` returns an empty array if it cannot resolve the provided
path, regardless of whether the provided path contains a globbed element
or not.

This commit reorganized the code in `compile_obj` and `compile_input_path`
to take into account the provided search paths during globbing.  As a
side-effect we always call `compile_input_path` with a fully resolved
path to the input file, and do not need to search the file in all search
paths again.

The CompileError for inputs which are not found has been moved to the
globbing phase, and is raised when `glob.glob` returns an empty array
for a given `input_path` in all provided search paths.