Add caching for fetched inventories and dependencies
Store the fetched items into .dependency_cache directory which is located in
the base direcotry or output-path directory.
When items are fetched, check if it exists in the cache first and then
safe-copy it to the output location.
When force fetched, overwrite the existing cache item with the newly
fetched items and recursively copy the items to the output location.