Make temporary directory for unpacking http(s) dependencies unique
When fetching dependencies with unpack=True using parallelism > 1 we
need to ensure that the temporary directory which is created to unpack
the dependency is unique to each multiprocessing worker.

This commit makes the temporary directory unique by appending the
multiprocessing worker process name to the directory name. Since the
temporary directory for unpacking the tarball is deleted after a single
dependency is fetched, using the worker process name is unique enough to
avoid the FileExistsError which we've previously sometimes observed when
two or more dependencies with unpack=True are fetched simultaneously.