Fix FileExistsError when fetching multiple dependencies with unpack=True (#608)
* Add test for parallel tar-ball dependency fetching

We've observed that dependency fetching with the default parallelism
setting of 4 and with multiple http(s) dependencies with unpack=True
sometimes fails with:

```
Unknown (Non-Kapitan) Error occurred
multiprocessing.pool.RemoteTraceback:
"""
Traceback (most recent call last):
  File "/usr/lib/python3.8/multiprocessing/pool.py", line 125, in worker
    result = (True, func(*args, **kwds))
  File "/path/to/venv/lib/python3.8/site-packages/kapitan/dependency_manager/base.py", line 176, in fetch_http_dependency
    os.makedirs(unpack_output)
  File "/usr/lib/python3.8/os.py", line 221, in makedirs
    mkdir(name, mode)
FileExistsError: [Errno 17] File exists: '/tmp/tmpj3f3nldp.kapitan/extracted'
"""
```

This commit adds a test which covers parallel dependency fetching for
http(s) dependencies with unpack=True.

Locally, the test has quite reliably replicated the issue.

* Make temporary directory for unpacking http(s) dependencies unique

When fetching dependencies with unpack=True using parallelism > 1 we
need to ensure that the temporary directory which is created to unpack
the dependency is unique to each multiprocessing worker.

This commit makes the temporary directory unique by appending the
multiprocessing worker process name to the directory name. Since the
temporary directory for unpacking the tarball is deleted after a single
dependency is fetched, using the worker process name is unique enough to
avoid the FileExistsError which we've previously sometimes observed when
two or more dependencies with unpack=True are fetched simultaneously.