Only propagate env vars which are present for input type `external`
I've noticed that the current implementation for propagating the
environment variables `$HOME` and `$PATH` to the subprocess spawned for
the external input type can raise an error if either or both of the
variables aren't present in Kapitan's environment.

When one of the variables was missing, the original implementation would
propagate value `None` to the subprocess, which causes a `TypeError`
when trying to spawn the subprocess:

```
Unknown (Non-Kapitan) error occurred:
multiprocessing.pool.RemoteTraceback:
"""
Traceback (most recent call last):
  File "/usr/lib/python3.8/multiprocessing/pool.py", line 125, in worker
    result = (True, func(*args, **kwds))
  File "/home/simon/work/syn/commodore/.tox/py38/lib/python3.8/site-packages/kapitan/targets.py", line 500, in compile_target
    input_compiler.compile_obj(comp_obj, ext_vars, **kwargs)
  File "/home/simon/work/syn/commodore/.tox/py38/lib/python3.8/site-packages/kapitan/inputs/base.py", line 57, in compile_obj
    self.compile_input_path(input_path, comp_obj, ext_vars, **kwargs)
  File "/home/simon/work/syn/commodore/.tox/py38/lib/python3.8/site-packages/kapitan/inputs/base.py", line 72, in compile_input_path
    self.compile_file(
  File "/home/simon/work/syn/commodore/.tox/py38/lib/python3.8/site-packages/kapitan/inputs/external.py", line 58, in compile_file
    env_vars = {k: compiled_target_pattern.sub(compile_path, v) for (k, v) in self.env_vars.items()}
  File "/home/simon/work/syn/commodore/.tox/py38/lib/python3.8/site-packages/kapitan/inputs/external.py", line 58, in <dictcomp>
    env_vars = {k: compiled_target_pattern.sub(compile_path, v) for (k, v) in self.env_vars.items()}
TypeError: expected string or bytes-like object
```

This commit updates the propagation logic to only propagate `$HOME` and
`$PATH` if they're actually present in Kapitan's environment.