Refactor handling of inventory backend command line flag
This commit switches the inventory backend command line flag to the form
`--inventory-backend=<backend name>`. Additionally, we restrict the
allowed values for the flag to the known backends.

To ensure the allowed values for the command line flag, and the known
implementations are in sync, we introduce a dict holding the mapping
from flag values to `Inventory` subclasses.

Finally, we ensure that the selected backend is cached as a string in
`cached.args["inventory-backend"]`. Note that this is not how
`cached.args` is used otherwise, but since it's unlikely that there ever
will be a real command `inventory-backend`, this should be fine.