Make normalisation of `nodes_uri` and `classes_uri` optional
This commit introduces a new optional parameter for
`get_reclass_config()` which allows callers to disable the normalisation
of the `nodes_uri` and `classes_uri` parameters.

This option is required for reclass-rs, which expects these parameters
to be relative to the inventory path, rather than relative to the
working directory.