Add test case for `componse-node-name`
This commit is based on the test case introduced in PR #1138.

We move the compose-node-name test implementation to `tests/` so it's
stored in the same place as the other tests and make the test class
inheritable so that we can reuse the test for reclass-rs.

Additionally we extend the test to check that all targets found by
`search_targets()` are rendered by `render_targets()`. This is needed
since we don't use reclass(-rs)'s target discovery logic in
`search_targets()` but instead use a simplified version that's
implemented directly in Kapitan's inventory backend base class.

However, `render_targets()` then renders whatever targets reclass(-rs)'s
target discovery finds, so if there's a mismatch we'd want to be
informed.

Co-authored-by: Matteo Voges <matteo.voges@nexenio.com>
Co-authored-by: Simon Gerber <simon.gerber@vshn.ch>